Create ERD Object

This function serves as a constructor for an Entity-Relationship Diagram (ERD) object. This object encapsulates both the data frames representing the entities and the relationships between these entities. The function takes as its arguments a list of data frames and a list of relationships and returns a list object of class "ERD".

Usage

create_erd(df_list, relationships)

Arguments

df_list

A named list of data frames, where each data frame represents an entity in the ERD. The names of the list elements correspond to the names of the entities.

relationships

A nested named list describing the relationships between entities. The top-level names in this list should correspond to the names in df_list. Each element of this list is itself a list, describing relationships that the corresponding entity has with other entities. The list of acceptable values is specified in "Details."

Value

An object of class "ERD", which is a named list containing two elements:

data_framesNamed list of data frames identical to df_list.

relationshipamed list of relationships identical to relationships.

The class attribute of this list is set to "ERD".

Details

Possible values in each relationship element of the list include:

"||" which indicates one and only one

">|" which indicates one or more (left table)

"|<" which indicates one or more (right table)

">0" which indicates zero or more (left table)

"0<" which indicates zero or more (right table)

"|0" which indicates zero or one (left table)

"0|" which indicates zero or one (right table)

It is imperative that the names used in df_list and relationships are consistent, as these are used for creating the ERD object and for subsequent operations like rendering and performing joins.

Users can effortlessly encapsulate the data and relationships pertaining to an ERD into a single R object with this function, thereby facilitating downstream operations like rendering and joining.

Examples

NULL
#> NULL